0.1.0
Dominio minimamente viável

Começo de integração com Kotlin

+++ DockIn.Backend/src/DockIn.Application/Dtos/StockDto.cs
@@ -3,7 +3,7 @@ namespace DockIn.Application.Dtos;
 public record StockDto(
     int StockId,
     int ArtigoId,
-    double Preco,
+    decimal Preco,
     int Quantidade,
     int QuantidadeReservada,
     DateTime CreatedAt
